:sparkles: split math lib into new file
