+++ src/nasal_builtin.cpp
@@ -243,64 +243,6 @@ var builtin_u32not(var* local, gc& ngc) {
     return var::num((f64)(u32)(~u32(local[1].num())));
 }
 
-var builtin_pow(var* local, gc& ngc) {
-    var x = local[1];
-    var y = local[2];
-    if (x.type!=vm_num || y.type!=vm_num) {
-        return var::num(std::nan(""));
-    }
-    return var::num(std::pow(x.num(), y.num()));
-}
-
-var builtin_sin(var* local, gc& ngc) {
-    var val = local[1];
-    return var::num(val.type==vm_num?sin(val.num()):std::nan(""));
-}
-
-var builtin_cos(var* local, gc& ngc) {
-    var val = local[1];
-    return var::num(val.type==vm_num?cos(val.num()):std::nan(""));
-}
-
-var builtin_tan(var* local, gc& ngc) {
-    var val = local[1];
-    return var::num(val.type==vm_num?tan(val.num()):std::nan(""));
-}
-
-var builtin_exp(var* local, gc& ngc) {
-    var val = local[1];
-    return var::num(val.type==vm_num?exp(val.num()):std::nan(""));
-}
-
-var builtin_lg(var* local, gc& ngc) {
-    var val = local[1];
-    return var::num(val.type==vm_num?log(val.num())/log(10.0):std::nan(""));
-}
-
-var builtin_ln(var* local, gc& ngc) {
-    var val = local[1];
-    return var::num(val.type==vm_num?log(val.num()):std::nan(""));
-}
-
-var builtin_sqrt(var* local, gc& ngc) {
-    var val = local[1];
-    return var::num(val.type==vm_num?sqrt(val.num()):std::nan(""));
-}
-
-var builtin_atan2(var* local, gc& ngc) {
-    var x = local[1];
-    var y = local[2];
-    if (x.type!=vm_num || y.type!=vm_num) {
-        return var::num(std::nan(""));
-    }
-    return var::num(atan2(y.num(), x.num()));
-}
-
-var builtin_isnan(var* local, gc& ngc) {
-    var x = local[1];
-    return (x.type==vm_num && std::isnan(x.num()))?one:zero;
-}
-
 var builtin_time(var* local, gc& ngc) {
     var val = local[1];
     if (val.type!=vm_num) {
@@ -1306,16 +1248,6 @@ nasal_builtin_table builtin[] = {
     {"__u32or", builtin_u32or},
     {"__u32nand", builtin_u32nand},
     {"__u32not", builtin_u32not},
-    {"__pow", builtin_pow},
-    {"__sin", builtin_sin},
-    {"__cos", builtin_cos},
-    {"__tan", builtin_tan},
-    {"__exp", builtin_exp},
-    {"__lg", builtin_lg},
-    {"__ln", builtin_ln},
-    {"__sqrt", builtin_sqrt},
-    {"__atan2", builtin_atan2},
-    {"__isnan", builtin_isnan},
     {"__time", builtin_time},
     {"__contains", builtin_contains},
     {"__delete", builtin_delete},
