+++ src/nasal_builtin.h
@@ -59,16 +59,6 @@ var builtin_u32and(var*, gc&);
 var builtin_u32or(var*, gc&);
 var builtin_u32nand(var*, gc&);
 var builtin_u32not(var*, gc&);
-var builtin_pow(var*, gc&);
-var builtin_sin(var*, gc&);
-var builtin_cos(var*, gc&);
-var builtin_tan(var*, gc&);
-var builtin_exp(var*, gc&);
-var builtin_lg(var*, gc&);
-var builtin_ln(var*, gc&);
-var builtin_sqrt(var*, gc&);
-var builtin_atan2(var*, gc&);
-var builtin_isnan(var*, gc&);
 var builtin_time(var*, gc&);
 var builtin_contains(var*, gc&);
 var builtin_delete(var*, gc&);
